Fixed the alert box classes. I intend to add an abstract class uniting both of these classes later since they have very similar functionality. I renamed the MazeNavView class to MazeNavUI to fit the naming conventions of my teammates' use cases. I made additional changes to some of the entities.
