--- src/main/java/screens/AlertBox.java
@@ -1,39 +1,0 @@
-package screens;
-
-import javafx.event.EventHandler;
-import javafx.geometry.Pos;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.layout.VBox;
-import javafx.stage.Modality;
-import javafx.stage.Stage;
-import javafx.event.ActionEvent;
-
-public abstract class AlertBox implements EventHandler<ActionEvent> {
-
-    public static void display(String title, String message) {
-        Stage window = new Stage();
-
-        window.initModality(Modality.APPLICATION_MODAL);
-        window.setTitle(title);
-        window.setMinWidth(250);
-
-        Label label = new Label();
-        label.setText(message);
-        Button closeButton = new Button("Close");
-        closeButton.setOnAction(EventHandler<ActionEvent>);
-
-        VBox layout = new VBox(10);
-        layout.getChildren().addAll(label, closeButton);
-        layout.setAlignment(Pos.CENTER);
-
-        Scene scene = new Scene(layout);
-        window.setScene(scene);
-        window.showAndWait();
-    }
-
-    public void handle(ActionEvent event) {
-
-    }
-}
