+++ src/main/java/screens/InvalidMoveAlertBox.java
@@ -1,11 +1,38 @@
 package screens;
 
-import java.awt.event.ActionEvent;
+import javafx.geometry.Pos;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.layout.VBox;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
 
-public class InvalidMoveAlertBox extends AlertBox { // These alert boxes could be separated if you were smart about it
+
+
+public class InvalidMoveAlertBox {
+    // These alert boxes could be separated if you were smart about it
     // Create another class called alert box and do it through there
 
-    public void handle(ActionEvent event) {
+    public static void display(String title, String message) {
+        Stage window = new Stage();
 
+        window.initModality(Modality.APPLICATION_MODAL);
+        window.setTitle(title);
+        window.setMinWidth(250);
+
+        Label label = new Label();
+        label.setText(message);
+        Button closeButton = new Button("Close");
+        closeButton.setOnAction(e -> window.close());
+
+        VBox layout = new VBox(10);
+        layout.getChildren().addAll(label, closeButton);
+        layout.setAlignment(Pos.CENTER);
+
+        Scene scene = new Scene(layout);
+        window.setScene(scene);
+        window.showAndWait();
     }
+
 }
