+++ src/main/java/screens/MazeCompleteAlertBox.java
@@ -1,12 +1,15 @@
 package screens;
 
-import javafx.stage.Stage;
-import javafx.stage.Modality;
-import javafx.scene.control.Label;
-import javafx.scene.control.Button;
-import javafx.scene.layout.VBox;
+import javafx.event.EventHandler;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.layout.VBox;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
+import javafx.event.ActionEvent;
+
 
 // Could you possibly combine the two classes InvalidMoveAlertBox and MazeCompleteAlertBox?
 public class MazeCompleteAlertBox {
@@ -21,7 +24,15 @@ public class MazeCompleteAlertBox {
         Label label = new Label();
         label.setText(message);
         Button closeButton = new Button("Close");
-        closeButton.setOnAction(e -> window.close()); // This needs to include changing the screen using scene manager
+        EventHandler<ActionEvent> popuphandle = new EventHandler<ActionEvent>() {
+            @Override
+            public void handle(ActionEvent actionEvent) {
+                ScreenManager.changeScreen("PlayerHomeScreen");
+                window.close();
+            }
+        };
+
+        closeButton.setOnAction(popuphandle);
 
         VBox layout = new VBox(10);
         layout.getChildren().addAll(label, closeButton);
@@ -30,6 +41,5 @@ public class MazeCompleteAlertBox {
         Scene scene = new Scene(layout);
         window.setScene(scene);
         window.showAndWait();
-        // Change the other screen here
     }
 }
